ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
replicate==1.0.4
